+++ projects/ngx-hateoas-client/src/lib/service/internal/resource-http.service.spec.ts
@@ -108,12 +108,12 @@ describe('ResourceHttpService', () => {
   });
 
   it('GET REQUEST should evict cache when returned object is not resource', () => {
-    spyOn(LibConfig, 'getConfig').and.returnValue({
+    spyOn(LibConfig, 'getConfig').and.returnValue(LibConfig.mergeConfigs({
       ...LibConfig.DEFAULT_CONFIG,
       cache: {
         enabled: true
       }
-    });
+    }));
     httpClientSpy.get.and.returnValue(of({any: 'value'}));
 
     resourceHttpService.get('someUrl').subscribe(() => {
@@ -124,12 +124,12 @@ describe('ResourceHttpService', () => {
 
   it('GET REQUEST should use cache when useCache param is TRUE', () => {
     httpClientSpy.get.and.returnValue(of(rawResource));
-    spyOn(LibConfig, 'getConfig').and.returnValue({
+    spyOn(LibConfig, 'getConfig').and.returnValue(LibConfig.mergeConfigs({
       ...LibConfig.DEFAULT_CONFIG,
       cache: {
         enabled: true
       }
-    });
+    }));
 
     resourceHttpService.get('order', {
       useCache: true
@@ -140,12 +140,12 @@ describe('ResourceHttpService', () => {
 
   it('GET REQUEST should NOT use cache when useCache param is FALSE', () => {
     httpClientSpy.get.and.returnValue(of(rawResource));
-    spyOn(LibConfig, 'getConfig').and.returnValue({
+    spyOn(LibConfig, 'getConfig').and.returnValue(LibConfig.mergeConfigs({
       ...LibConfig.DEFAULT_CONFIG,
       cache: {
         enabled: true
       }
-    });
+    }));
 
     resourceHttpService.get('order', {
       useCache: false